+++ packaging/core.re.template
@@ -362,7 +362,7 @@ acSetChkFilePathPerm {msiSetChkFilePathPerm("disallowPathReg"); }
 acSetVaultPathPolicy {msiSetGraftPathScheme("no","1"); }
 #
 # 17) acSetReServerNumProc - This rule set the policy for the number of processes
-# to use when running jobs in the irodsReServer. The irodsReServer can now
+# to use when running jobs in the irodsDelayServer. The irodsDelayServer can now
 # multi-task such that one or two long running jobs cannot block the execution
 # of other jobs. One function can be called:
 #    msiSetReServerNumProc(numProc) - numProc can be "default" or a number
