+++ server/delay_server/src/irodsDelayServer.cpp
@@ -1,4 +1,4 @@
-#include "irods/irodsReServer.hpp"
+#include "irods/irodsDelayServer.hpp"
 
 #include "irods/catalog.hpp"
 #include "irods/client_connection.hpp"
